Fix: Bundle Lit library to prevent breaking other cards
ROOT CAUSE:
The AutoSnooze card was loading Lit from external CDNs, which caused
cascade failures in Lovelace's resource loading system. When the CDN
import failed or loaded asynchronously, it would break other custom
cards (Mushroom cards, auto-entities, etc.) resulting in "Custom
element doesn't exist" errors across the dashboard.

SOLUTION:
Following the same pattern as popular cards like Mushroom, the card
now bundles Lit library directly into the output file using Rollup.
This eliminates all external dependencies and prevents the card from
breaking other resources.

Changes:
- Added build system with Rollup and terser
- Created src/autosnooze-card.js with regular lit imports
- Build bundles Lit into a single 45KB self-contained file
- No external CDN dependencies
- Auto-registration still enabled (safe now with bundled dependencies)
- Added package.json with lit@2.8.0 and rollup build scripts
- Added .gitignore for node_modules

Version: 2.7.0

This approach matches industry best practices for Home Assistant custom
cards and ensures AutoSnooze will never break other dashboard cards.
